Many enhancements: - Added more efficient decimation for low CPU usage at 2MSPS - Many tweaks - Code cleanup
git-svn-id: https://subversion.assembla.com/svn/sdrsharp/trunk@228 d3cde5d6-82b8-4669-8b62-43054d6781d8

+++ PanView/SpectrumAnalyzer.cs
@@ -283,7 +283,7 @@ namespace SDRSharp.PanView
             float bandpassOffset;
             var bandpassWidth = 0f;
             var cursorWidth = Math.Max((_filterBandwidth + _filterOffset) * _xIncrement, 2);
-            var xCarrier = ClientRectangle.Width / 2 + (_frequency - _displayCenterFrequency) * _xIncrement;
+            var xCarrier = (float) ClientRectangle.Width / 2 + (_frequency - _displayCenterFrequency) * _xIncrement;
 
             switch (_bandType)
             {
